backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
sqlalchemy
python-dotenv